// src/generated-code/angular/horseless-odata-content-workspace/horselesscontent/microsoft/graph/passwordcredential.complex.ts
//#region ODataApiGen ODataImports
import {
  Duration
} from 'angular-odata';//#endregion

//#region ODataApiGen Imports
//#endregion

export interface PasswordCredential {
  //#region ODataApiGen Properties
  customKeyIdentifier?: ArrayBuffer;
  displayName?: string;
  endDateTime?: Date;
  keyId?: string;
  startDateTime?: Date;
  secretText?: string;
  hint?: string;
  //#endregion
}